Entity property cleanup (#977)
* move spawn point to EntityManager

* set spawn points from world model load

* move temporary PickupPlayer property out of entity

* move armor properties to player

* use subsector id, change slow tick properties to shorts

* remove virtual ApplyArmorDamage, correctly check for KillDamage before modifying

+++ Core/World/Entities/Enemy.cs
@@ -319,7 +319,7 @@ public partial class Entity
             MoveCount = WorldStatic.Random.NextByte() & 15;
 
         if (WorldStatic.SlowTickEnabled)
-            ChaseFailureSkipCount = WorldStatic.SlowTickChaseFailureSkipCount + (ChaseFailureCount++ & 1);
+            ChaseFailureSkipCount = (short)(WorldStatic.SlowTickChaseFailureSkipCount + (ChaseFailureCount++ & 1));
 
         // Need to try to use the monster's normal movement speed if stuck. Otherwise they may never move or correctly cross teleport lines.
         ClosetChaseSpeed = MonsterMovementSpeed;
